Add UI class
