+++ src/main/java/duke/UI/UI.java
@@ -1,0 +1,7 @@
+package duke.UI;
+
+public class UI {
+    public void showIntro() {
+        System.out.println("\tHi handsome! My name is Duck. What can I do for you?");
+    }
+}
